rename: v1 채팅 레포지토리 디렉토리 변경
