+++ chat/src/main/java/com/team15gijo/chat/domain/repository/v1/ChatMessageRepository.java
@@ -1,4 +1,4 @@
-package com.team15gijo.chat.domain.repository;
+package com.team15gijo.chat.domain.repository.v1;
 
 import com.team15gijo.chat.domain.model.ChatMessageDocument;
 import org.springframework.data.mongodb.repository.MongoRepository;
